creating parse-uri module using parse-uri in ajax instead of creating anchor elements adding test for parse-uri
